backend/requirements.txt
flask
flask-cors
opencv-python
insightface
onnxruntime
faiss-cpu
numpy
PyJWT
werkzeug

